requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
colorama==0.4.6
exceptiongroup==1.0.4
h11==0.14.0
idna==3.4
iniconfig==1.1.1
numpy==1.23.5
opencv-python==4.1.2.30
outcome==1.2.0
packaging==21.3
Pillow==9.3.0
pluggy==1.0.0
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.2.0
PyYAML==6.0
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
wsproto==1.2.0
